refactor(ClanConfig): extract symbol selection logic into SymbolSelector component
The symbol selection logic was duplicated and tightly coupled within the ClanConfig component. By extracting it into a dedicated SymbolSelector component, we improve code maintainability and reusability. This change also follows the Single Responsibility Principle, making the codebase easier to understand and modify in the future.
